+++ muckpot-api/src/main/kotlin/com/yapp/muckpot/common/redisson/DistributedLockAop.kt
@@ -1,5 +1,6 @@
 package com.yapp.muckpot.common.redisson
 
+import com.yapp.muckpot.redis.RedissonCallNewTransaction
 import mu.KLogging
 import org.aspectj.lang.ProceedingJoinPoint
 import org.aspectj.lang.annotation.Around
@@ -41,7 +42,7 @@ class DistributedLockAop(
                 rLock.unlock()
             } catch (e: IllegalMonitorStateException) {
                 // 이미 unlock된 상태에서 다시 unlock 시도시 발생, 에러로그로 남김
-                log.error(e.stackTrace.joinToString("\n"))
+                log.error(e) { "" }
             }
         }
     }
